Доработки контроллеров и методов typeorm для users и wishes

+++ src/wishes/wishes.service.ts
@@ -1,4 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  NotFoundException,
+  Injectable,
+  Logger,
+} from '@nestjs/common';
 import { CreateWishDto } from './dto/create-wish.dto';
 import { UpdateWishDto } from './dto/update-wish.dto';
 
@@ -11,25 +16,69 @@ export class WishesService {
   constructor(
     @InjectRepository(Wish)
     private wishesRepository: Repository<Wish>,
-  ) {}
+  ) {
+    this.logger = new Logger(WishesService.name);
+  }
+  private readonly logger: Logger;
 
-  async create(createWishDto: CreateWishDto) {
-    return await this.wishesRepository.create(createWishDto);
+  async create(createWishDto: CreateWishDto, userId) {
+    return await this.wishesRepository.insert({
+      ...createWishDto,
+      owner: userId,
+    });
   }
 
-  async findAll() {
-    return await this.wishesRepository.find();
+  async findWishBiId(id: number) {
+    return await this.wishesRepository.findOne({
+      where: {
+        id,
+      },
+      relations: {
+        owner: true,
+        offers: true,
+      },
+    });
   }
 
-  async findOne(id: number) {
-    return await this.wishesRepository.findOneByOrFail({ id });
-  }
+  async updateWishById(
+    id: number,
+    updateWishDto: UpdateWishDto,
+    userId: number,
+  ) {
+    const wish = await this.findWishBiId(id);
 
-  async update(id: number, updateWishDto: UpdateWishDto) {
+    if (!wish) {
+      throw new NotFoundException('Такого подарка не существует');
+    }
+    if (wish.owner.id !== userId) {
+      throw new ForbiddenException('Вы не можете редактировать чужие подарки');
+    }
+    //TODO проверить!!!
+    if (wish.raised > 0 && updateWishDto.price) {
+      throw new ForbiddenException(
+        'Вы не можете редактировать цену, если уже есть желающие скинуться',
+      );
+    }
+
     return await this.wishesRepository.update(id, updateWishDto);
   }
 
-  async remove(id: number) {
+  async deleteWishById(id: number, userId: number) {
+    const wish = await this.findWishBiId(id);
+
+    if (!wish) {
+      throw new NotFoundException('Такого подарка не существует');
+    }
+    if (wish.owner.id !== userId) {
+      throw new ForbiddenException('Вы не можете удалять чужие подарки');
+    }
+
     await this.wishesRepository.delete(id);
+    return wish;
+  }
+
+  // //////////////////////////////
+  async findAll() {
+    return await this.wishesRepository.find();
   }
 }
